# Speed up conversion to StaticArray