--- a/src/convert.jl
+++ b/src/convert.jl
@@ -13,9 +13,13 @@
     unroll_tuple(a, Length(a))
 end
 
+@noinline function dimension_mismatch_fail(SA::Type, a::AbstractArray)
+    error("Dimension mismatch. Expected input array of length $(length(SA)), got length $(length(a))")
+end
+
 @inline function convert(::Type{SA}, a::AbstractArray) where {SA <: StaticArray}
     @boundscheck if length(a) != length(SA)
-        error("Dimension mismatch. Expected input array of length $(length(SA)), got length $(length(a))")
+        dimension_mismatch_fail(SA, a)
     end
 
     return SA(unroll_tuple(a, Length(SA)))