@timholy 2017-09-24
Actually I should confess that I was the one who introduced this annoying but useful trick to speed up a couple of (long-standing) hindrances to performance that I happened to notice while I was doing my inlining work. In this case it's not the throw that's the problem, it's the GC root that gets created even when that branch doesn't get taken (because of the string interpolation). If the `throw` were the problem, the difference between these versions would be vastly larger. In an ideal world the `@noinline` annotation wouldn't be necessary, but the inlining cost model currently doesn't account for GC roots. (It runs long before the GC pass is taken, and it would be pretty nontrivial to figure this stuff out.)

Indeed, there's a [special case](https://github.com/JuliaLang/julia/blob/2e864155d0498e73c35f92c91efb4b4799affea1/base/inference.jl#L4771-L4776) that gives `throw` statements a cost estimate of 0, since they really shouldn't ever be part of the normal control flow in a function and hence represent a branch that doesn't affect normal execution.