@c42f 2017-09-24
Ah I see. I was thinking that - as a special case - any function such as `error` which contains an unconditional `throw` should never be inlined. If nothing else, to save a bit on code size and icache pressure. But it's very interesting that the problem is in the GC root setup which happens in the frame where the message is formatted.