@c42f 2017-11-04
I just noticed https://github.com/JuliaLang/julia/pull/24431 - it looks like we won't need these kind of workarounds fairly soon :-)